backend/requirements.txt
fastapi
uvicorn
langchain==0.2.14
langchain-community==0.2.12
langchain-openai
faiss-cpu
python-dotenv
tiktoken
langchain-text-splitters
